end tour when clicked on connect apps button and tooltip text change
